[FEATURE] Creacion de los serviciosy controladores de login

+++ src/main/java/com/tfg/swapCatBack/integration/adapters/mappers/HistoryAdapterMapper.java
@@ -45,7 +45,7 @@ public class HistoryAdapterMapper implements AdapterMapper<HistoryInfoDTO>{
     private HistoryInfoDTO maper(JsonNode node) {
         return HistoryInfoDTO.builder()
                 .priceUsd(node.get(1).asDouble())
-                .time(node.get(0).asDouble())
+                .time(node.get(0).asLong())
                 .build();
     }
 }
